bugs fixed

+++ src/Co_Evolution_Manager/main.java
@@ -4,9 +4,9 @@ import java.io.File;
 import java.util.Scanner;
 
 public class main {
-	
+
 	public static Scanner scanner;
-	
+
 	public static void main (String[] args) {
 
 		/*	Allowed strategies
@@ -14,10 +14,10 @@ public class main {
 		 * "nsyncsourceNignorelocal"
 		 * */		
 		scanner = new Scanner(System.in);		
-		new configure ("sa.nt", "sd.nt", "ta.nt", "td.nt", "t.nt", "syncsourceNkeeplocalWresolvedconflicts", "NT");   
+		new configure ("sa", "sd", "ta", "td", "t", "syncsourceNignorelocal"/*"syncsourceNkeeplocalWresolvedconflicts"*/, "RDF/XML");   
 		strategy.apply ();
 		emptyResources ();
-		renameOutput ();
+		renameOutput ("t");
 		//Conflict_Resolver.statistics.findBlankNodes("rdf:Description"); 
 		scanner.close();
 	}
@@ -25,35 +25,44 @@ public class main {
 
 	public static void emptyResources () {
 		try {
-			File f = new File (configure.initialTarget);
-			f.delete();
+			File f;
 
-			f = new File (configure.sourceAdditionsChangeset);
-			f.delete();
-
-			f = new File (configure.sourceDeletionsChangeset);
-			f.delete();
-
-			f = new File (configure.targetAdditionsChangeset);
-			f.delete();
-
-			f = new File (configure.targetDeletionsChangeset);
-			f.delete();			
+			if (configure.initialTarget != null) {
+				f = new File (configure.initialTarget);
+				f.delete();
+			}
+			if (configure.sourceAdditionsChangeset != null) {			
+				f = new File (configure.sourceAdditionsChangeset);
+				f.delete();
+			}
+			if (configure.sourceDeletionsChangeset != null) {
+				f = new File (configure.sourceDeletionsChangeset);
+				f.delete();
+			}
+			if (configure.targetAdditionsChangeset != null) {
+				f = new File (configure.targetAdditionsChangeset);
+				f.delete();
+			}
+			if (configure.targetDeletionsChangeset != null) {
+				f = new File (configure.targetDeletionsChangeset);
+				f.delete();			
+			}
 		} catch(Exception e){  		
 			e.printStackTrace();
 		}
 	}
 
-	
-	public static void renameOutput () 	{
-		
-		File ifile = new File(configure.newTarget);
-		File ofile = new File(configure.initialTarget);
+	// Output target will act as initial input for next time synchronization, so rename it at end of synchronization
+	public static void renameOutput (String outputfilename) 	{
+		if (configure.newTarget != null) {
+			File ifile = new File(configure.newTarget);
+			File ofile = new File(outputfilename);
 
-		if (ofile.exists()){
-				System.out.println("file exists");
+			if (ofile.exists()){
+				System.out.println("file already exists");
 			} 
-		ifile.renameTo(ofile);
+			ifile.renameTo(ofile);
+		}
 	}
 
 }
